Move empty songs in a different dir

+++ lyricsFileNameReprocessor.ts
@@ -2,9 +2,10 @@ import fs from 'fs';
 import path from 'path';
 import dotenv from 'dotenv';
 import * as process from 'process';
-import { TXT_EXTENSION } from './constants';
-import { normalizeFileName } from './src';
+import { EMPTY_STRING, TXT_EXTENSION } from './constants';
+import { normalizeFileName, SongSection } from './src';
 import recursive from 'recursive-readdir';
+import _ from 'lodash';
 
 dotenv.config();
 
@@ -16,23 +17,32 @@ const reprocessFileNames = async (dir: string) => {
     .forEach((filePath) => {
       const existingContent = fs.readFileSync(filePath).toString();
 
-      const fileName = path.basename(filePath);
-      const newFileName = normalizeFileName(fileName);
-      console.log(
-        `Processing "${fileName}": ${
-          fileName !== newFileName ? newFileName : 'No change.'
-        }.`,
-      );
+      const strings = existingContent
+        .replaceAll('^(\r)\n', '\r\n')
+        // Replace to .split(/\n\n/gim) in test
+        .split(/\r\n\r\n/gim)
+        .filter(Boolean)
+        .map(_.trim);
+      console.log(strings);
+      if (strings.length === 1) {
+        const fileName = path.basename(filePath);
+        // const newFileName = normalizeFileName(fileName);
+        // console.log(
+        //   `Processing "${fileName}": ${
+        //     fileName !== newFileName ? newFileName : 'No change.'
+        //   }.`,
+        // );
 
-      fs.unlinkSync(filePath);
-      fs.writeFileSync(
-        path.join(__dirname, path.dirname(filePath), newFileName),
-        existingContent,
-      );
+        fs.unlinkSync(filePath);
+        fs.writeFileSync(
+          path.join(__dirname, 'candidates-to-lookup', fileName),
+          existingContent,
+        );
+      }
     });
 };
 
 (async () => {
-  await reprocessFileNames(process.env.CANDIDATES_DIR);
-  await reprocessFileNames(process.env.VERIFIED_DIR);
+  await reprocessFileNames('./candidates-upcoming');
+  // await reprocessFileNames(process.env.VERIFIED_DIR);
 })();
